refactor(nvimage): Extract NRRD image reader to ImageReaders module
