+++ packages/niivue/src/nvimage/index.ts
@@ -542,7 +542,10 @@ export class NVImage {
         break
       case NVIMAGE_TYPE.NHDR:
       case NVIMAGE_TYPE.NRRD:
-        imgRaw = await newImg.readNRRD(dataBuffer as ArrayBuffer, pairedImgData) // detached
+        imgRaw = await ImageReaders.Nrrd.readNrrd(newImg, dataBuffer as ArrayBuffer)
+        if (imgRaw === null) {
+          throw new Error(`Failed to parse NHDR/NRRD file ${name}`)
+        }
         break
       case NVIMAGE_TYPE.MHD:
       case NVIMAGE_TYPE.MHA:
@@ -2230,303 +2233,6 @@ export class NVImage {
     return [outVs, v1s] */
     return outVs
   } // readMIF()
-
-  // not included in public docs
-  // read NRRD format image
-  // http://teem.sourceforge.net/nrrd/format.html
-  async readNRRD(dataBuffer: ArrayBuffer, pairedImgData: ArrayBuffer | null): Promise<ArrayBuffer> {
-    // inspired by parserNRRD.js in https://github.com/xtk
-    // Copyright (c) 2012 The X Toolkit Developers <dev@goXTK.com>
-    // http://www.opensource.org/licenses/mit-license.php
-    this.hdr = new NIFTI1()
-    const hdr = this.hdr
-    hdr.pixDims = [1, 1, 1, 1, 1, 0, 0, 0]
-    const len = dataBuffer.byteLength
-    // extract initial text header
-    let txt = null
-    const bytes = new Uint8Array(dataBuffer)
-    for (let i = 1; i < len; i++) {
-      if (bytes[i - 1] === 10 && bytes[i] === 10) {
-        const v = dataBuffer.slice(0, i - 1)
-        txt = new TextDecoder().decode(v)
-        hdr.vox_offset = i + 1
-        break
-      }
-    }
-
-    if (txt === null) {
-      throw new Error('could not extract txt')
-    }
-
-    const lines = txt.split('\n')
-    if (!lines[0].startsWith('NRRD')) {
-      alert('Invalid NRRD image')
-    }
-    const n = lines.length
-    let isGz = false
-    let isMicron = false
-    let isDetached = false
-    const mat33 = mat3.fromValues(NaN, 0, 0, 0, 1, 0, 0, 0, 1)
-    const offset = vec3.fromValues(0, 0, 0)
-    let rot33 = mat3.create()
-    for (let i = 1; i < n; i++) {
-      let str = lines[i]
-      if (str[0] === '#') {
-        continue
-      } // comment
-      str = str.toLowerCase()
-      const items = str.split(':')
-      if (items.length < 2) {
-        continue
-      }
-      const key = items[0].trim()
-      let value = items[1].trim()
-      value = value.replaceAll(')', ' ')
-      value = value.replaceAll('(', ' ')
-      value = value.trim()
-      switch (key) {
-        case 'data file':
-          isDetached = true
-          break
-        case 'encoding':
-          if (value.includes('raw')) {
-            isGz = false
-          } else if (value.includes('gz')) {
-            isGz = true
-          } else {
-            alert('Unsupported NRRD encoding')
-          }
-          break
-        case 'type':
-          switch (value) {
-            case 'uchar':
-            case 'unsigned char':
-            case 'uint8':
-            case 'uint8_t':
-              hdr.numBitsPerVoxel = 8
-              hdr.datatypeCode = NiiDataType.DT_UINT8
-              break
-            case 'signed char':
-            case 'int8':
-            case 'int8_t':
-              hdr.numBitsPerVoxel = 8
-              hdr.datatypeCode = NiiDataType.DT_INT8
-              break
-            case 'short':
-            case 'short int':
-            case 'signed short':
-            case 'signed short int':
-            case 'int16':
-            case 'int16_t':
-              hdr.numBitsPerVoxel = 16
-              hdr.datatypeCode = NiiDataType.DT_INT16
-              break
-            case 'ushort':
-            case 'unsigned short':
-            case 'unsigned short int':
-            case 'uint16':
-            case 'uint16_t':
-              hdr.numBitsPerVoxel = 16
-              hdr.datatypeCode = NiiDataType.DT_UINT16
-              break
-            case 'int':
-            case 'signed int':
-            case 'int32':
-            case 'int32_t':
-              hdr.numBitsPerVoxel = 32
-              hdr.datatypeCode = NiiDataType.DT_INT32
-              break
-            case 'uint':
-            case 'unsigned int':
-            case 'uint32':
-            case 'uint32_t':
-              hdr.numBitsPerVoxel = 32
-              hdr.datatypeCode = NiiDataType.DT_UINT32
-              break
-            case 'float':
-              hdr.numBitsPerVoxel = 32
-              hdr.datatypeCode = NiiDataType.DT_FLOAT32
-              break
-            case 'double':
-              hdr.numBitsPerVoxel = 64
-              hdr.datatypeCode = NiiDataType.DT_FLOAT64
-              break
-            default:
-              throw new Error('Unsupported NRRD data type: ' + value)
-          }
-          break
-        case 'spacings':
-          {
-            const values = value.split(/[ ,]+/)
-            for (let d = 0; d < values.length; d++) {
-              hdr.pixDims[d + 1] = parseFloat(values[d])
-            }
-          }
-          break
-        case 'sizes':
-          {
-            const dims = value.split(/[ ,]+/)
-            hdr.dims[0] = dims.length
-            for (let d = 0; d < dims.length; d++) {
-              hdr.dims[d + 1] = parseInt(dims[d])
-            }
-          }
-          break
-        case 'endian':
-          if (value.includes('little')) {
-            hdr.littleEndian = true
-          } else if (value.includes('big')) {
-            hdr.littleEndian = false
-          }
-          break
-        case 'space directions':
-          {
-            const vs = value.split(/[ ,]+/)
-            if (vs.length !== 9) {
-              break
-            }
-            for (let d = 0; d < 9; d++) {
-              mat33[d] = parseFloat(vs[d])
-            }
-          }
-          break
-        case 'space origin':
-          {
-            const ts = value.split(/[ ,]+/)
-            if (ts.length !== 3) {
-              break
-            }
-            offset[0] = parseFloat(ts[0])
-            offset[1] = parseFloat(ts[1])
-            offset[2] = parseFloat(ts[2])
-          }
-          break
-        case 'space units':
-          if (value.includes('microns')) {
-            isMicron = true
-          }
-          break
-        case 'space':
-          if (value.includes('right-anterior-superior') || value.includes('ras')) {
-            rot33 = mat3.fromValues(
-              1,
-              0,
-              0,
-
-              0,
-              1,
-              0,
-
-              0,
-              0,
-              1
-            )
-          } else if (value.includes('left-anterior-superior') || value.includes('las')) {
-            rot33 = mat3.fromValues(
-              -1,
-              0,
-              0,
-
-              0,
-              1,
-              0,
-
-              0,
-              0,
-              1
-            )
-          } else if (value.includes('left-posterior-superior') || value.includes('lps')) {
-            rot33 = mat3.fromValues(
-              -1,
-              0,
-              0,
-
-              0,
-              -1,
-              0,
-
-              0,
-              0,
-              1
-            )
-          } else {
-            log.warn('Unsupported NRRD space value:', value)
-          }
-          break
-        default:
-          log.warn('Unknown:', key)
-      } // read line
-    } // read all lines
-    if (!isNaN(mat33[0])) {
-      // if spatial transform provided
-      this.hdr.sform_code = 2
-      if (isMicron) {
-        // convert micron to mm
-        // @ts-expect-error FIXME: converting mat3 to mat4
-        mat4.multiplyScalar(mat33, mat33, 0.001)
-        offset[0] *= 0.001
-        offset[1] *= 0.001
-        offset[2] *= 0.001
-      }
-      if (rot33[0] < 0) {
-        offset[0] = -offset[0]
-      } // origin L<->R
-      if (rot33[4] < 0) {
-        offset[1] = -offset[1]
-      } // origin A<->P
-      if (rot33[8] < 0) {
-        offset[2] = -offset[2]
-      } // origin S<->I
-      mat3.multiply(mat33, rot33, mat33)
-      const mat = mat4.fromValues(
-        mat33[0],
-        mat33[3],
-        mat33[6],
-        offset[0],
-        mat33[1],
-        mat33[4],
-        mat33[7],
-        offset[1],
-        mat33[2],
-        mat33[5],
-        mat33[8],
-        offset[2],
-        0,
-        0,
-        0,
-        1
-      )
-      const mm000 = this.vox2mm([0, 0, 0], mat)
-      const mm100 = this.vox2mm([1, 0, 0], mat)
-      vec3.subtract(mm100, mm100, mm000)
-      const mm010 = this.vox2mm([0, 1, 0], mat)
-      vec3.subtract(mm010, mm010, mm000)
-      const mm001 = this.vox2mm([0, 0, 1], mat)
-      vec3.subtract(mm001, mm001, mm000)
-      hdr.pixDims[1] = vec3.length(mm100)
-      hdr.pixDims[2] = vec3.length(mm010)
-      hdr.pixDims[3] = vec3.length(mm001)
-      hdr.affine = [
-        [mat[0], mat[1], mat[2], mat[3]],
-        [mat[4], mat[5], mat[6], mat[7]],
-        [mat[8], mat[9], mat[10], mat[11]],
-        [0, 0, 0, 1]
-      ]
-    }
-
-    if (isDetached && pairedImgData) {
-      // ??? .gz files automatically decompressed?
-      return pairedImgData.slice(0)
-    }
-    if (isDetached) {
-      log.warn('Missing data: NRRD header describes detached data file but only one URL provided')
-    }
-    if (isGz) {
-      return await NVUtilities.decompressToBuffer(new Uint8Array(dataBuffer.slice(hdr.vox_offset)))
-    } else {
-      return dataBuffer.slice(hdr.vox_offset)
-    }
-  } // readNRRD()
 
   // not included in public docs
   // Transform to orient NIfTI image to Left->Right,Posterior->Anterior,Inferior->Superior (48 possible permutations)
